+++ src/modules/mouseHover/store/actionsMouseHover.js
@@ -109,9 +109,18 @@ export default {
             return layer?.mouseHoverField && layer.mouseHoverField !== "";
         }));
     },
+    /**
+     * Highlights the given feeature depending on the geometry type by dispatching "Maps/highlightFeature".
+     * @param {Object} context the context
+     * @param {Object} context.dispatch the dispatch
+     * @param {Object} context.state the state
+     * @param {Object} payload the state
+     * @param {Object} payload.feature openlayers feature to hioghlight
+     * @param {Object} payload.layer openlayers layer og the feature
+     * @returns {void}
+     */
     highlightFeature ({dispatch, state}, {feature, layer}) {
         const {highlightVectorRulesPointLine, highlightVectorRulesPolygon} = state,
-            layerId = layer.get("id"),
             featureGeometryType = feature.getGeometry().getType(),
             featureId = feature.getId(),
             styleObj = featureGeometryType.toLowerCase().indexOf("polygon") > -1 ?
@@ -124,18 +133,17 @@ export default {
                 feature: feature,
                 scale: styleObj.image?.scale
             },
-            rawLayer = rawLayerList.getLayerWhere({id: layerId});
+            rawLayer = rawLayerList.getLayerWhere({id: layer.get("id")});
 
         if (featureGeometryType === "LineString" || featureGeometryType === "MultiLineString") {
             highlightObject.type = "highlightLine";
         }
-        layer.id = layerId;
         highlightObject.zoomLevel = styleObj.zoomLevel;
         if (rawLayer && rawLayer.styleId) {
             highlightObject.styleId = rawLayer.styleId;
         }
-        else if (layer && layer.styleId) {
-            highlightObject.styleId = layer.styleId;
+        else if (layer.get("styleId")) {
+            highlightObject.styleId = layer.get("styleId");
         }
 
         highlightObject.highlightStyle = {