+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 ### Fixed
 - Issue \#1624: Fixed misleading parameter in config.js.md
+- Issue \#1625: Fixed using configured styleId when mousehover on vectorlayers.
 - FileImport: importDrawLayer reappears in the layer tree when importing a file, even after being removed.
 - SearchBar: Fixed misaligned action icons in "Show all" results for street searches.